// v4/migrations/snapshots/a015335828566b81776b520ee0ac05c783aaf5a7f3735de716df603a260ccccc/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Album: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly parentId: CodecTypes['pg/uuid@1']['output'] | null;
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly path: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly body: CodecTypes['pg/text@1']['output'];
      readonly visibility: 'public' | 'hidden' | 'private';
      readonly isOpenForSubalbums: CodecTypes['pg/bool@1']['output'];
      readonly eventMetadataUrl: CodecTypes['pg/text@1']['output'];
      readonly ordering: CodecTypes['pg/int4@1']['output'];
      readonly eventDate: CodecTypes['pg/date-string@1']['output'];
      readonly isDownloadable: CodecTypes['pg/bool@1']['output'];
      readonly termsId: CodecTypes['pg/uuid@1']['output'] | null;
      readonly ownerId: CodecTypes['pg/uuid@1']['output'] | null;
      readonly thumbnailPhotoId: CodecTypes['pg/uuid@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly AlbumCredit: {
      readonly albumId: CodecTypes['pg/uuid@1']['output'];
      readonly photographerId: CodecTypes['pg/uuid@1']['output'];
      readonly isCopyright: CodecTypes['pg/bool@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly ordering: CodecTypes['pg/int4@1']['output'];
    };
    readonly Media: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly photoId: CodecTypes['pg/uuid@1']['output'];
      readonly role: 'original' | 'preview' | 'thumbnail';
      readonly format: 'jpeg' | 'webp' | 'avif' | 'heif';
      readonly width: CodecTypes['pg/int4@1']['output'];
      readonly height: CodecTypes['pg/int4@1']['output'];
      readonly storageKey: CodecTypes['pg/text@1']['output'];
      readonly byteSize: CodecTypes['pg/int4@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly Photo: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly albumId: CodecTypes['pg/uuid@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly path: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly ordering: CodecTypes['pg/int4@1']['output'];
      readonly takenAt: CodecTypes['pg/timestamptz-string@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly Photographer: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly displayName: CodecTypes['pg/text@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly introduction: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'] | null;
      readonly defaultTermsId: CodecTypes['pg/uuid@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly PhotographerLink: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly photographerId: CodecTypes['pg/uuid@1']['output'];
      readonly href: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly ordering: CodecTypes['pg/int4@1']['output'];
    };
    readonly Terms: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly text: CodecTypes['pg/text@1']['output'];
      readonly url: CodecTypes['pg/text@1']['output'];
      readonly ownerId: CodecTypes['pg/uuid@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly sub: CodecTypes['pg/text@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly displayName: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Album: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly parentId: CodecTypes['pg/uuid@1']['input'] | null;
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly path: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly body: CodecTypes['pg/text@1']['input'];
      readonly visibility: 'public' | 'hidden' | 'private';
      readonly isOpenForSubalbums: CodecTypes['pg/bool@1']['input'];
      readonly eventMetadataUrl: CodecTypes['pg/text@1']['input'];
      readonly ordering: CodecTypes['pg/int4@1']['input'];
      readonly eventDate: CodecTypes['pg/date-string@1']['input'];
      readonly isDownloadable: CodecTypes['pg/bool@1']['input'];
      readonly termsId: CodecTypes['pg/uuid@1']['input'] | null;
      readonly ownerId: CodecTypes['pg/uuid@1']['input'] | null;
      readonly thumbnailPhotoId: CodecTypes['pg/uuid@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly AlbumCredit: {
      readonly albumId: CodecTypes['pg/uuid@1']['input'];
      readonly photographerId: CodecTypes['pg/uuid@1']['input'];
      readonly isCopyright: CodecTypes['pg/bool@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly ordering: CodecTypes['pg/int4@1']['input'];
    };
    readonly Media: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly photoId: CodecTypes['pg/uuid@1']['input'];
      readonly role: 'original' | 'preview' | 'thumbnail';
      readonly format: 'jpeg' | 'webp' | 'avif' | 'heif';
      readonly width: CodecTypes['pg/int4@1']['input'];
      readonly height: CodecTypes['pg/int4@1']['input'];
      readonly storageKey: CodecTypes['pg/text@1']['input'];
      readonly byteSize: CodecTypes['pg/int4@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly Photo: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly albumId: CodecTypes['pg/uuid@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly path: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly ordering: CodecTypes['pg/int4@1']['input'];
      readonly takenAt: CodecTypes['pg/timestamptz-string@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly Photographer: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly displayName: CodecTypes['pg/text@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly introduction: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'] | null;
      readonly defaultTermsId: CodecTypes['pg/uuid@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly PhotographerLink: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly photographerId: CodecTypes['pg/uuid@1']['input'];
      readonly href: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly ordering: CodecTypes['pg/int4@1']['input'];
    };
    readonly Terms: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly text: CodecTypes['pg/text@1']['input'];
      readonly url: CodecTypes['pg/text@1']['input'];
      readonly ownerId: CodecTypes['pg/uuid@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly sub: CodecTypes['pg/text@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly displayName: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;